requirements.txt
Flask
google-cloud-vision
gunicorn
Werkzeug
google-cloud-aiplatform